Moved enums and exceptions to separate files
