+++ exceptions/exceptions.py
@@ -1,0 +1,8 @@
+from utils.models import User
+
+
+class NeedsNewTokens(Exception):
+    def __init__(self, user: User, access_token: str, refresh_token: str):
+        self.user = user
+        self.access_token = access_token
+        self.refresh_token = refresh_token
