+++ utils/enums.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+class ValidPermissions(Enum):
+    DELETE_ORGANIZATION = "Delete Organization"
+    EDIT_ORGANIZATION = "Edit Organization"
+    INVITE_USER = "Invite User"
+    REMOVE_USER = "Remove User"
+    EDIT_USER_ROLE = "Edit User Role"
+    CREATE_ROLE = "Create Role"
+    DELETE_ROLE = "Delete Role"
+    EDIT_ROLE = "Edit Role"
